remove user image function in setting-ui and add user image ui into user files, Additionally, add object class of image model
