+++ app/src/main/java/org/maternalcare/modules/main/settings/ui/SettingsUI.kt
@@ -1,13 +1,8 @@
 package org.maternalcare.modules.main.settings.ui
 
-import android.net.Uri
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -17,15 +12,12 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.IconButtonDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -35,31 +27,22 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
-import coil.compose.AsyncImage
 import kotlinx.coroutines.launch
 import org.maternalcare.R
 import org.maternalcare.modules.intro.IntroNav
 import org.maternalcare.modules.intro.login.viewmodel.UserState
 import org.maternalcare.modules.main.MainNav
-
-@Preview(showSystemUi = true)
-@Composable
-fun SettingsUIPreview() {
-    SettingsUI(navController = rememberNavController())
-}
+import org.maternalcare.modules.main.user.model.dto.UserDto
+import org.maternalcare.modules.main.user.ui.UserImageUI
 
 @Composable
-fun SettingsUI(navController: NavController) {
+fun SettingsUI(navController: NavController, currentUserDto: UserDto) {
     val vm = UserState.current
     val coroutineScope = rememberCoroutineScope()
     var showButton by remember { mutableStateOf(true) }
@@ -72,7 +55,7 @@ fun SettingsUI(navController: NavController) {
         horizontalAlignment = Alignment.CenterHorizontally
 
     ) {
-        Profile(navController)
+        Profile(navController, currentUserDto)
         Spacer(
             modifier = Modifier
                 .height(30.dp)
@@ -83,6 +66,7 @@ fun SettingsUI(navController: NavController) {
                     coroutineScope.launch {
                         showButton = false
                         vm.signOut()
+
                         navController.navigate(IntroNav.Login) {
                             popUpTo(0)
                         }
@@ -126,20 +110,18 @@ fun SettingsUI(navController: NavController) {
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun Profile(navController: NavController){
-    var picUrl by remember { mutableStateOf("") }
-    val saveImageUri: (Uri) -> Unit = { uri -> picUrl = uri.toString() }
+fun Profile(navController: NavController, currentUserDto: UserDto){
     Column(
         modifier = Modifier
             .fillMaxWidth()
             .padding(16.dp),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        ImagePicker(selectedImageUri = Uri.parse(picUrl), onImageSelected = saveImageUri)
+        UserImageUI()
 
         Spacer(modifier = Modifier.padding(vertical = 15.dp))
 
-        UserDetails()
+        UserDetails(currentUserDto)
 
         Spacer(modifier = Modifier.padding(vertical = 5.dp))
 
@@ -148,93 +130,29 @@ fun Profile(navController: NavController){
 }
 
 @Composable
-fun UserDetails() {
-    val userDetails = listOf( "Juan", "R.", "Dela Cruz" )
+fun UserDetails(currentUserDto: UserDto) {
+    val userDetails = listOf(
+        currentUserDto.firstName,
+        currentUserDto.middleName,
+        currentUserDto.lastName
+    )
     Row(
         modifier = Modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.Center
     ) {
         userDetails.forEach { fullName ->
-            Text(
-                text = fullName,
-                fontSize = 17.sp,
-                fontFamily = FontFamily.SansSerif,
-                color = Color(0xFF6650a4),
-                modifier = Modifier.padding(horizontal = 3.dp)
-            )
-        }
-    }
-}
-
-@Composable
-fun ImagePicker(selectedImageUri: Uri? = null, onImageSelected: (Uri) -> Unit = {}) {
-    var selectedImgUri by remember { mutableStateOf<Uri?>(null) }
-    val photoPickerLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.PickVisualMedia(),
-        onResult = { uri ->
-            uri?.let {
-                selectedImgUri = it
-                onImageSelected(it)
-            }
-        }
-    )
-
-    Box(
-        Modifier.height(140.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .size(140.dp)
-                .clip(CircleShape)
-                .background(Color(0xFF6650a4))
-                .border(3.dp, Color(0xFF6650a4), CircleShape),
-            contentAlignment = Alignment.Center
-        ) {
-            if (selectedImgUri != null) {
-                AsyncImage(
-                    modifier = Modifier
-                        .clip(CircleShape)
-                        .size(140.dp)
-                        .clip(CircleShape),
-                    model = selectedImgUri,
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop
-                )
-            } else {
-                Icon(
-                    painter = painterResource(id = R.drawable.person),
-                    contentDescription = "Default placeholder",
-                    modifier = Modifier
-                        .size(90.dp),
-                    tint = Color.White,
-
+            if (fullName != null) {
+                Text(
+                    text = fullName,
+                    fontSize = 17.sp,
+                    fontFamily = FontFamily.SansSerif,
+                    color = Color(0xFF6650a4),
+                    modifier = Modifier.padding(horizontal = 3.dp)
                 )
             }
         }
-        IconButton(
-            onClick = {
-                photoPickerLauncher.launch(
-                PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
-            },
-            modifier = Modifier
-                .size(35.dp)
-                .padding(2.dp)
-                .clip(CircleShape)
-                .align(Alignment.BottomEnd),
-            colors = IconButtonDefaults.iconButtonColors(Color(0xFF6650a4)),
-        ) {
-            Icon(
-                painter = painterResource(id = R.drawable.cameraalt),
-                contentDescription = null,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(5.dp),
-                tint = Color.White
-            )
-        }
     }
 }
-
 
 @Composable
 fun Setting(navController: NavController){
