+++ app/src/main/java/org/maternalcare/modules/main/user/ui/UserImageUI.kt
@@ -1,0 +1,100 @@
+package org.maternalcare.modules.main.user.ui
+
+import android.net.Uri
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.PickVisualMediaRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.IconButtonDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import org.maternalcare.R
+
+@Composable
+fun UserImageUI(selectedImageUri: Uri? = null, onImageSelected: (Uri) -> Unit = {}) {
+    var selectedImgUri by remember { mutableStateOf<Uri?>(null) }
+    val photoPickerLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.PickVisualMedia(),
+        onResult = { uri ->
+            uri?.let {
+                selectedImgUri = it
+                onImageSelected(it)
+            }
+        }
+    )
+
+    Box(
+        Modifier.height(140.dp)
+    ) {
+        Box(
+            modifier = Modifier
+                .size(140.dp)
+                .clip(CircleShape)
+                .background(Color(0xFF6650a4))
+                .border(3.dp, Color(0xFF6650a4), CircleShape),
+            contentAlignment = Alignment.Center
+        ) {
+            if (selectedImgUri != null) {
+                AsyncImage(
+                    modifier = Modifier
+                        .clip(CircleShape)
+                        .size(140.dp)
+                        .clip(CircleShape),
+                    model = selectedImgUri,
+                    contentDescription = null,
+                    contentScale = ContentScale.Crop
+                )
+            } else {
+                Icon(
+                    painter = painterResource(id = R.drawable.person),
+                    contentDescription = "Default placeholder",
+                    modifier = Modifier
+                        .size(90.dp),
+                    tint = Color.White,
+
+                    )
+            }
+        }
+        IconButton(
+            onClick = {
+                photoPickerLauncher.launch(
+                    PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
+            },
+            modifier = Modifier
+                .size(35.dp)
+                .padding(2.dp)
+                .clip(CircleShape)
+                .align(Alignment.BottomEnd),
+            colors = IconButtonDefaults.iconButtonColors(Color(0xFF6650a4)),
+        ) {
+            Icon(
+                painter = painterResource(id = R.drawable.cameraalt),
+                contentDescription = null,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(5.dp),
+                tint = Color.White
+            )
+        }
+    }
+}
